fix: resolve circular dependacy
