+++ package.json
@@ -20,20 +20,20 @@
     "@ethersproject/strings": "5.6.1",
     "@ethersproject/wallet": "5.6.2",
     "@gnosis.pm/safe-apps-web3-react": "0.6.8",
-    "@honeycomb-finance/airdrop": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/bridge": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/core": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/elixir": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/governance": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/honeycomb-provider": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/pools": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/portfolio": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/sar": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/shared": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/state-hooks": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/swap": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/wallet-connectors": "0.0.1-internal-20230922074227",
-    "@honeycomb-finance/walletmodal": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/airdrop": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/bridge": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/core": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/elixir": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/governance": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/honeycomb-provider": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/pools": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/portfolio": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/sar": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/shared": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/state-hooks": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/swap": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/wallet-connectors": "0.0.1-internal-20230928072245",
+    "@honeycomb-finance/walletmodal": "0.0.1-internal-20230928072245",
     "@metamask/jazzicon": "2.0.0",
     "@pangolindex/exchange-contracts": "2.1.7",
     "@pangolindex/governance": "1.2.0",
@@ -125,7 +125,7 @@
   },
   "scripts": {
     "start": "vite",
-    "build": "tsc && vite build",
+    "build": "set NODE_OPTIONS=--max-old-space-size=558193 && tsc && vite build",
     "serve": "serve build -l 3000",
     "version": "npm run build",
     "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --fix",
